refactor: move runtime coverage options to engine
