+++ crates/engine/src/lib.rs
@@ -66,7 +66,7 @@ pub use fallow_core::duplicates::{
 pub use fallow_types::discover::{DiscoveredFile, FileId};
 pub use fallow_types::extract::ModuleInfo;
 pub use fallow_types::results::AnalysisResults;
-pub use health::{HealthAnalysisResult, HealthSort};
+pub use health::{HealthAnalysisResult, HealthSort, RuntimeCoverageOptions};
 
 /// Result alias for typed engine operations.
 pub type EngineResult<T> = Result<T, EngineError>;
